+++ ExilePartyClient/src/app/authorize/authorize.module.ts
@@ -5,12 +5,13 @@ import { MatListModule } from '@angular/material/list';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { RouterModule } from '@angular/router';
 
+import { IncomeService } from '../shared/providers/income.service';
 import { PartyService } from '../shared/providers/party.service';
 import { SharedModule } from '../shared/shared.module';
 import { AuthorizeComponent } from './authorize.component';
-import { PartyModule } from './party/party.module';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { InspectPlayersModule } from './inspect-players/inspect-players.module';
+import { PartyModule } from './party/party.module';
 
 @NgModule({
   imports: [
@@ -29,6 +30,9 @@ import { InspectPlayersModule } from './inspect-players/inspect-players.module';
     MatSidenavModule
   ],
   declarations: [AuthorizeComponent],
-  providers: [PartyService]
+  providers: [
+    PartyService,
+    IncomeService
+  ]
 })
 export class AuthorizeModule { }