+++ ExilePartyClient/src/app/shared/interfaces/chart.interface.ts
@@ -1,0 +1,10 @@
+export interface ChartSeries {
+  name: string;
+  series: ChartSeriesEntry[];
+}
+
+export interface ChartSeriesEntry {
+  name: string | Date;
+  value: number;
+  items: any[];
+}