+++ ExilePartyClient/src/app/shared/interfaces/player.interface.ts
@@ -1,4 +1,5 @@
 import { Character } from './character.interface';
+import { NetWorthSnapshot } from './income.interface';
 
 export interface Player {
     connectionID: string;
@@ -11,10 +12,9 @@ export interface Player {
     inArea: Array<string>;
     generic: boolean;
     genericHost: string;
+    netWorthSnapshots: NetWorthSnapshot[];
 }
 
 export interface RecentPlayer {
     name: string;
-    invited: boolean;
-    private: boolean;
 }